backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# HTTP client for integrations
httpx==0.26.0
aiohttp==3.9.1

# Image processing
Pillow==10.2.0

# AWS/Azure SDK (for cloud services)
boto3==1.34.25

# Security
cryptography==41.0.7

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.0
isort==5.13.2
mypy==1.8.0

# PDF generation for audit packets
reportlab==4.0.8
PyPDF2==3.0.1

# Date/time handling
python-dateutil==2.8.2

# Rate limiting
slowapi==0.1.9
